First working version for Sparkle support.
Added reference to Sparkle framework to the project. The framework was **NOT** included in the project folder, so as not to "pollute" the source directory. Also, the framework is quite big. Right now, it is referenced as an absolute path to '/Library/Frameworks/Sparkle.framework', which means developers will have to upload it separately and put it there. Maybe we should change that?

Also updated the Main.nib to include an instance of SUUpdater.

Needed: add item in Preferences. Test that everything works as expected. This will only make sense once 0.1.2 is released with Sparkle and a version 0.1.3 is ready.